Game Loop
Project was originally made in an individual file and not via github, so I had to copy my project into the new file, sorry about that.

This is the simple game loop. Including a superman and car model, a win/loss trigger at the car y-level detecting the fallSpeed of superman as he lands, as well as a system for calculating deceleration and aceleration speed using addition and subtraction.

When the player presses the space key, supermans fallSpeed will gradually decrease until finally being negative which allows superman to fly upwards.

I also added a limit to supermans fallSpeed so he won't fall too fast. Then I added an invisible "roof" so that superman cannot fly outside the canvas.