Startscreen and Layout
Edited the structure of the code. Added a "state" variable that can be used to define if the user is in the start screen or playing the game.

Added a simple start screen with the text "press any button to play", if any key is pressed then the "state" variable switches to "game" and the another screen for the game comes up.